+++ components/hashes/blake2/src/blake2s.rs
@@ -4,10 +4,11 @@ use utils::Hash;
 
 use super::Blake2;
 
+#[derive(Debug)]
 pub struct Blake2s(Blake2<u32>);
 
 impl Blake2s {
-    pub fn with_digest_len(n: usize) -> Self {
+    pub fn with_digest_len(n: u8) -> Self {
         Self(Blake2::<u32>::with_digest_len(n))
     }
 }
